--- index.html
@@ -1,72 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>My portfolio</title>
-    <link rel="stylesheet" href="./css/main.css">
-</head>
-<body>
-    <div class="header">
-        <h1 class="header-title">Palkina Sofia</h1>
-        <p class="header-subtitle">frontend developer</p>
-    </div>
-
-    <div class="portfolio">
-        <div class="container">
-            <h2 class="portfolio-header">Portfolio</h2>
-            <div class="portfolio-cards">
-                <div class="card">
-                    <a href="https://github.com/angel-palkina/Portfolio_site">
-                        <img class="card-image" src="./images/portfolio/project1.jpg" alt="First website">
-                    </a>
-                    <h3 class="card-title">First website</h3>
-                </div>
-                <div class="card">
-                    <a href="https://github.com/angel-palkina/SSAD">
-                        <img class="card-image" src="./images/portfolio/project2.jpg" alt="First C++ program">
-                    </a>
-                    <h3 class="card-title">First C++ program</h3>
-                    
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <div class="footer">
-        <div class="container">
-            <div class="footer-row">
-                <button class="comic">
-                    <a href="./comic.html" class="comic-ref">The XKCD comic</a>
-                </button>
-                <div class="about-me">
-                    <ul class="bio">
-                        <li>Place of study: Innopolis University</li>
-                        <li>Hometown: Perm</li>
-                        <li>Date of Birth: 29.03.2005</li>
-                        <li>Secondary general education: Lyceum 10 Perm</li>
-
-                    </ul>
-                </div>
-                <div class="footer-coperate">
-                    <div class="footer-coporate-name">© Palkina Sofia </div>
-                    <div class="footer-icons">
-                        <a href="https://vk.com/angel.sofia" id="vk-link">
-                            <img class="footer-image" src="./images/footer/vk.png" alt="vk">
-                        </a>
-                        <a href="https://github.com/angel-palkina" id="gh-link">
-                            <img class="footer-image" src="./images/footer/git.png" alt="git">
-                        </a>
-                        <a href="https://t.me/angel_sofia2005" id="tg-link">
-                            <img class="footer-image" src="./images/footer/tg.png" alt="tg">
-                        </a>
-                        <a href="mailto:s.palkina@innopolis.university" class="footer-buttom-mail" id="email">
-                            <img class="footer-image" src="./images/footer/mail.png" alt="mail">
-                        </a>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-</body>
-</html>
