+++ src/css/main.css
@@ -1,0 +1,162 @@
+*{
+    box-sizing: border-box;
+}
+
+body{
+    margin: 0;
+    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+}
+
+.container {
+    max-width: 1100px ;
+    margin-left: auto;
+    margin-right: auto;
+
+}
+
+/*Header*/
+.header {
+    width: 100%;
+    height: 100vh;
+    background-color: #7ac6c2;
+    background-image: url(./../../images/header/header-back.jpg);
+    background-size: cover;
+    background-position: center;
+
+    /* выравнивание */
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
+.header-title {
+    margin: 0;
+
+    font-weight: bold;
+    font-size: 56px;
+    line-height: 130%;
+
+    text-align: center;
+
+    color: #ffffff;
+}
+.header-subtitle {
+    margin: 0;
+
+    font-weight: bold;
+    font-size: 16px;
+    line-height: 150%;
+
+    text-align: center;
+    letter-spacing: 0.2cm;
+    text-transform: uppercase;
+
+    color: #ffffff;
+
+}
+
+/*portfolio*/
+
+.portfolio {
+    background-color: #f7f6f6;
+    padding-top: 80px;
+    padding-bottom: 115px;
+}
+
+.portfolio-header {
+    font-weight: 700;
+    font-size: 46px;
+    line-height: 130%;
+    text-align: center;
+    color: #000;
+}
+.portfolio-cards{
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-between;
+}
+.card {
+    background-color: #ffffff ;
+    max-width: 540px;
+    padding: 20px;
+}
+.card-image {
+    margin-bottom: 20px;
+    width: 100%;
+    height: 80%;
+}
+.card-title {
+    margin: 0;
+    margin-bottom: 13px;
+    font-weight: 600;
+    font-size: 24px;
+    line-height: 130%;
+    color: #000;
+}
+
+/*Footer*/
+
+.footer {
+    background-color: #7ac6c2;
+    padding-top: 30px ;
+    padding-bottom: 20px;
+    width: 100%;
+    color: #000;
+}
+
+.footer-row {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+}
+.bio {
+    line-height: 130%;
+}
+.comic {
+    display: inline-block;
+    height: 50px;
+    padding-left: 30px;
+    padding-right: 30px;
+
+    margin-top: 15px;
+
+    border: 3px solid #000;
+
+    border-radius: 50px;
+
+    line-height: 44px;
+
+    background-color: #7ac6c2;
+
+}
+
+.comic-ref {
+    color: #000;
+    font-weight: 800;
+}
+.footer-coporate-name {
+    font-size: 28px;
+    font-weight: 700;
+    line-height: 130%;
+
+}
+.footer-icons {
+    display: flex;
+    align-items: center;
+    margin-top: 10px;
+}
+.footer-image {
+    max-width: 20px;
+    max-height: 20px;
+}
+.footer-icons > *{
+    margin-left: 30px;
+}
+.comic-picture {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    margin-top: 50px;
+    margin-bottom: 20px;
+}
